--- workspaces/jenkins/plugins/jenkins-scaffolder/src/actions/index.ts
@@ -1,1 +1,0 @@
-export * from './job-create';
