--- workspaces/jenkins/plugins/jenkins-scaffolder/src/actions/job-create/index.ts
@@ -1,3 +1,0 @@
-import { jenkinsJobCreate } from './module';
-
-export default jenkinsJobCreate;
