+++ workspaces/jenkins/plugins/jenkins-scaffolder/src/index.ts
@@ -1,8 +1,24 @@
-/***/
-/**
- * The jenkins-scaffolder module for @backstage/plugin-scaffolder-backend.
- *
- * @packageDocumentation
- */
+import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node/alpha';
+import { createBackendModule, coreServices } from '@backstage/backend-plugin-api';
+import { jenkinsCreateJobAction } from './action';
 
-export * from './actions';
+/**
+ * A backend module that registers the action into the scaffolder
+ */
+export const jenkinsJobCreate = createBackendModule({
+  pluginId: 'scaffolder',
+  moduleId: 'jenkins:job:create',
+  register(env) {
+    env.registerInit({
+      deps: {
+        scaffolder: scaffolderActionsExtensionPoint,
+        config: coreServices.rootConfig
+      },
+      async init({ scaffolder, config }) {
+        scaffolder.addActions(jenkinsCreateJobAction({config: config}))
+      },
+    });
+  },
+})
+
+export { jenkinsJobCreate as default };
